requirements.txt
argparse
requests
python-dotenv
rich
paramiko
cryptography